Don't allow EOS until 4 frames have been generated (#14761)
* Don't allow EOS until 4 frames have been generated

The number of frames is configurable via a parameter to infer_batch().

This is a workaround to the observation that when CFG is on generation sometimes terminates after zero tokens. That appears to be an artifact of CFG, since the EOS logits (both cond and uncon) are not particularly large  before CFG but become the maximum logit after it, with CFG amplifying the differences between the two.

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Formatting

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Command line option to set minimum number of frames to generate

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* formatting

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Show extreme values in violin plots

(to aid in debugging rare issues)

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Fix merge issues

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* More merge fixes

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Remove temporary changes in infer_and_evaluate.py

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Comments

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Comments

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

* Fix typo

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>

---------

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>